Changed the event a/c to that in script

Signed-off-by: shirisha Ganta <shiganta@in.ibm.com>