@shirishaganta 2020-07-21
Before Patch:
------------------
[root@ltczz256-lp3 perf]# avocado run perf_duplicate_probe.py 
/usr/local/lib/python3.6/site-packages/avocado_framework-80.0-py3.6.egg/avocado/plugins/run.py:285: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast and keep-tmp. 
  FutureWarning)
JOB ID     : 58b0d4308d74b1a106f4712cef9ba0a7fca76e61
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-07-20T23.56-58b0d43/job.log
 (1/1) perf_duplicate_probe.py:PerfProbe.test_probe: ERROR: Command 'perf probe select_task_rq_fair:15' failed.\nstdout: b''\nstderr: b"Probe point 'select_task_rq_fair:15' not found.\n  Error: Failed to add events.\n"\nadditional_info: None (127.51 s)
RESULTS    : PASS 0 | ERROR 1 | FAIL 0 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-07-20T23.56-58b0d43/results.html
JOB TIME   : 127.64 s
[root@ltczz256-lp3 perf]# 

After Patch:
--------------
[root@ltczz256-lp3 perf]# avocado run perf_duplicate_probe.py
/usr/local/lib/python3.6/site-packages/avocado_framework-80.0-py3.6.egg/avocado/plugins/run.py:285: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast and keep-tmp. 
  FutureWarning)
JOB ID     : 4f358eb1611fb80f165fcd26169732045713807d
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-07-21T01.29-4f358eb/job.log
 (1/1) perf_duplicate_probe.py:PerfProbe.test_probe: FAIL: perf probe is placing multiple probe at the same location  (2.28 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-07-21T01.29-4f358eb/results.html
JOB TIME   : 2.39 s
[root@ltczz256-lp3 perf]#