@shirishaganta 2020-07-21
After Patch:
[root@ltczz256-lp3 perf]# avocado run perf_duplicate_probe.py
/usr/local/lib/python3.6/site-packages/avocado_framework-80.0-py3.6.egg/avocado/plugins/run.py:285: FutureWarning: The following arguments will be changed to boolean soon: sysinfo, output-check, failfast and keep-tmp. 
  FutureWarning)
JOB ID     : 6da459aadc798a1768fb04763d3095713e71c5ef
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-07-21T02.20-6da459a/job.log
 (1/1) perf_duplicate_probe.py:PerfProbe.test_probe: FAIL: perf probe is placing multiple probe at the same location  (2.31 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-07-21T02.20-6da459a/results.html
JOB TIME   : 2.43 s
[root@ltczz256-lp3 perf]#