@shirishaganta 2020-07-27
In recent Rhel release, many perf related changes went in and the probe point 15 was worked earlier and now because of this changes we changed the probe point to 0 in order to work successfully