@PraveenPenguin 2020-07-27
> In recent Rhel release, many perf related changes went in and the probe point 15 was worked earlier and now because of this changes we changed the probe point to 0 in order to work successfully

@shirishaganta @harish-24 as this change from kernel so why not we control probe from YAML that time we need to change python code as changing python code frequent not a good practice and I have the same thought for other PR as well