@shirishaganta 2020-10-13
> > In recent Rhel release, many perf related changes went in and the probe point 15 was worked earlier and now because of this changes we changed the probe point to 0 in order to work successfully
> 
> @shirishaganta @harish-24 as this change from kernel so why not we control probe from YAML that time we need not change python code as changing python code frequent not a good practice and I have the same thought for other PR as well

@PraveenPenguin @harish-24 perf event is 0 for both rhel and suse. And also this probe point will not change frequently. So I dont think we need cfg file for this