@shirishaganta 2020-10-15
[root@ltcfleet8-lp2 perf]# avocado run perf_duplicate_probe.py
JOB ID     : ab4df1bc94e0e422e95491502ff5097284191d2b
JOB LOG    : /home/avocado-fvt-wrapper/results/job-2020-10-15T02.58-ab4df1b/job.log
 (1/1) perf_duplicate_probe.py:PerfProbe.test_probe: FAIL: perf probe is placing multiple probe at the same location  (1.61 s)
RESULTS    : PASS 0 | ERROR 0 | FAIL 1 | SKIP 0 | WARN 0 | INTERRUPT 0 | CANCEL 0
JOB HTML   : /home/avocado-fvt-wrapper/results/job-2020-10-15T02.58-ab4df1b/results.html
JOB TIME   : 2.45 s
[root@ltcfleet8-lp2 perf]#